Minor Alerations
I was just looking through the code and we need to do alot still. Still
setting up the 'settings' and 'help' page if we do need a help page. I
also replaced the time signature with the 34 picture. [Bjorn] I removed
the function from pressing the gear to tapping I thin you did that on
purpose though..